--- src/main/java/org/noostak/auth/application/GoogleService.java
@@ -1,5 +1,0 @@
-package org.noostak.auth.application;
-
-public interface GoogleService extends OauthService {
-
-}
