+++ src/main/java/org/noostak/auth/application/google/GoogleService.java
@@ -1,0 +1,7 @@
+package org.noostak.auth.application.google;
+
+import org.noostak.auth.application.OauthService;
+
+public interface GoogleService extends OauthService {
+
+}
